refactor: test config 한 클래스에 몰아서 사용
